fix: Move incorrectly used validator to Pydantic model
