+++ app/models/models.py
@@ -1,4 +1,3 @@
-from pydantic import validator
 from sqlalchemy import (
     Column,
     Integer,
@@ -47,12 +46,6 @@ class Todo(Base):
     )
     user = relationship("User", back_populates="todos")
 
-    @validator("title")
-    def title_must_not_be_empty(cls, v):
-        if not v:
-            raise ValueError("Title must not be empty")
-        return v
-
 
 class Habit(Base):
     __tablename__ = "habit"
@@ -76,28 +69,12 @@ class Habit(Base):
     user = relationship("User", back_populates="habits")
     habit_logs = relationship("HabitLog", back_populates="habit")
 
-    @validator("title")
-    def title_must_not_be_empty(cls, v):
-        if not v:
-            raise ValueError("Title must not be empty")
-        return v
+    @staticmethod
+    def repeat_days_to_string(repeat_days):
+        return "".join([str(day) for day in repeat_days])
 
-    @validator("repeat_days")
-    def validate_repeat_days(cls, v):
-        if len(v) == 0:
-            raise ValueError("repeat_days must not be empty")
-        for day in v:
-            if day not in range(1, 8):
-                raise ValueError(
-                    "repeat_days must be a list of integers between 1 and 7"
-                )
-        return v
-
-    @validator("start_time_minutes")
-    def validate_start_time_minutes(cls, v):
-        if v < 0 or v >= 1440:
-            raise ValueError("start_time_minutes must be between 0 and 1439")
-        return v
+    def repeat_days_to_list(self):
+        return [int(day) for day in self.repeat_days]
 
 
 class HabitLog(Base):
@@ -135,51 +112,6 @@ class Routine(Base):
         back_populates="routine",
         order_by="RoutineElement.order.asc()",
     )
-
-    @validator("repeat_days")
-    def validate_repeat_days(cls, v):
-        if len(v) == 0:
-            raise ValueError("repeat_days must not be empty")
-        for day in v:
-            if day not in range(1, 8):
-                raise ValueError(
-                    "repeat_days must be a list of integers between 1 and 7"
-                )
-        return v
-
-    @validator("start_time_minutes")
-    def validate_start_time_minutes(cls, v):
-        if v < 0 or v >= 1440:
-            raise ValueError("start_time_minutes must be between 0 and 1439")
-        return v
-
-    @validator("routine_elements")
-    def validate_routine_elements(cls, v):
-        for item in v:
-            if item.duration_minutes < 0:
-                raise ValueError("duration_minutes must be positive")
-            elif item.duration_minutes > 1440:
-                raise ValueError("duration_minutes must be less than 1440")
-        return v
-
-    @validator("routine_elements")
-    def validate_routine_elements_order(cls, v):
-        order_list = []
-        for item in v:
-            order_list.append(item.order)
-        order_list.sort()
-        for i in range(len(order_list)):
-            if order_list[i] != i + 1:
-                raise ValueError("routine_elements order must be 1, 2, 3, ...")
-        return v
-
-    @validator("title")
-    def validate_title(cls, v):
-        if len(v) == 0:
-            raise ValueError("title must not be empty")
-        elif len(v) > 255:
-            raise ValueError("title must be less than 255 characters")
-        return v
 
     def repeat_days_to_list(self):
         return [int(day) for day in self.repeat_days]
